sync/pyproject.toml
[tool.poetry]
name = "slack-emoji-leaderboard-sync"
version = "0.0.0"
description = ""
authors = ["Milo Weinberg <iapetus011@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.1.0"
aiohttp = "^3.8.3"
python-dotenv = "^0.21.0"
pydantic = "^1.10.4"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.11.4"
black = "^22.12.0"
autoflake = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
